feat: added consent, introduction, experiment description text and integration
