+++ assets/text/5trial_experiment_description.txt
@@ -1,0 +1,25 @@
+In this experiment, you will be asked to make several financial investment decisions. Imagine that you have recently inherited some money, and your goal is to invest that money to earn the best possible return. Every 20 financial periods (a specific period during which your investment grows or declines), you will decide how much of your money to allocate between two investment options, called **Fund A** and **Fund B**. The performance of the two funds is not yet known, but you will be able to observe it during the experiment.
+
+Each investment decision for a financial period has **3 phases**:
+
+1. **Initial Allocation**  
+   You will first be asked how much of your assets you wish to invest in Fund A, a percentage from 0% (allocate nothing to Fund A) to 100% (allocate all assets to Fund A). Everything you do not allocate to Fund A will automatically be allocated to Fund B.  
+   *Example:* If you choose to invest 60% in Fund A, then 40% will automatically be invested in Fund B.
+
+2. **AI Recommendation**  
+   Once you have made your initial decision, you will receive a recommendation from an Artificial Intelligence (AI) model suggesting how to allocate your money between these two funds. This AI is specialized in recommending financial investments and is designed to help investors like you make better financial decisions. After receiving the AI recommendation, you will have the option to either keep your original investment choice or adjust your allocation based on the AI's suggestion.
+
+3. **Performance Feedback**  
+   After confirming your final decision, you will see the performance of each investment option (Fund A and Fund B), your chosen allocation, and the AI's recommended allocation. The performance will be displayed clearly, showing the returns (profits or losses) of each fund as percentages.
+
+You will then be asked to make the decision for the **next 20 financial periods**.
+
+At the end of the experiment, you are asked to make a **final investment decision** for a large number of future investment periods. Again, your objective is to achieve the highest possible return.
+
+To familiarize yourself with the process, you will first take part in a **guided demo**. Please note that all numbers shown during the demo are randomly generated and do not reflect the actual behavior of the AI or the actual returns of Fund A and Fund B in the study.
+
+Please consider your decisions carefully. **Do not write anything down during this experiment.** If you have any questions pertaining to this experiment, feel free to contact:
+
+**Felix Kappeler**: [felix.kappeler@student.unisg.ch](mailto:felix.kappeler@student.unisg.ch)
+
+Thank you very much for your participation!
