--- assets/text/experiment_description.txt
@@ -1,26 +1,0 @@
-## Investment Decision Experiment
-
-This experiment simulates a series of investment decisions. In this experiment, imagine that you are an investment manager. In every financial period, you must decide how to split your assets between two investments, which we label **Fund A** and **Fund B**.
-
-You will be asked how much you wish to allocate to Fund A, an integer from **0** (allocate nothing to Fund A) to **100** (allocate all assets to Fund A). Whatever you do not allocate to Fund A will be automatically allocated to Fund B.
-
-After each decision, you will receive advice on the optimal asset allocation from our AI model, which is specially trained to recommend financial investments. You then have the opportunity to change your investment decision or continue with your initial one.
-
-After taking your final decision, information about the performance of:
-- **Fund A**
-- **Fund B**
-- **The AI’s recommendation**
-- **Your decision**
-
-will be displayed on the screen as four bars, each showing the percentage return of the respective investment strategy.
-
-You will then be asked to make the decision for the next financial period. At the end of the experiment, you will be asked to make one final investment decision for a large number of future periods. 
-
-Your responses in this experiment are very important to us. Please consider your decisions carefully and **do not write anything down during this experiment**.
-
-**To make sure you have read all the instructions, leave the "Place of Birth" field blank in the demographic questionnaire on the next page.**
-
-
-If you have any questions, please feel free to ask the experimenter at any time.
-
-Thank you for your help!
