+++ modules/subpages/consent.py
@@ -1,4 +1,5 @@
 import streamlit as st
+import os
 from modules.database import supabase
 from modules.session import update_session_progress
 
@@ -7,26 +8,30 @@ def show_consent():
 
     # Container with introduction & consent text
     with st.container():
-        intro_file_path = "assets/text/introduction.txt"
-        consent_file_path = "assets/text/consent.txt"
-
-        # Load intro text
-        with open(intro_file_path, "r", encoding="utf-8") as i:
-            intro_text = i.read()
+        # Load and display introduction text
+        intro_file_path = os.path.join("assets", "text", "introduction.txt")
+        intro_text = load_text(intro_file_path)
         st.markdown(intro_text)
         st.markdown("---")
 
-        # Load consent text
-        with open(consent_file_path, "r", encoding="utf-8") as f:
-            consent_text = f.read()
-        st.markdown(consent_text)
+        # Load and display consent text
+        consent_file_path = os.path.join("assets", "text", "consent.txt")
+        consent_text = load_text(consent_file_path)
+        st.markdown(consent_text)   
+
+        # Display the clickable text that expands to show more information
+        with st.expander("Obtain more information about the processing of your personal data"):
+            data_processing_file_path = os.path.join("assets", "text", "data_processing.txt")
+            data_processing_text = load_text(data_processing_file_path)
+            st.markdown(data_processing_text)
+
 
         # Consent form
         with st.form(key="consent_form"):
             consent_given = st.checkbox(
                 "I agree to the consent form and to the processing of my personal data in accordance with the information provided herein."
             )
-            submitted = st.form_submit_button("Agree & Continue")
+            submitted = st.form_submit_button("Continue")
             if submitted:
                 if consent_given:
                     # Update sessions table
@@ -39,3 +44,7 @@ def show_consent():
                     st.rerun()
                 else:
                     st.error("You must agree to participate to continue.")
+
+def load_text(file_path):
+    with open(file_path, "r", encoding="utf-8") as file:
+        return file.read()
