+++ modules/subpages/intro.py
@@ -1,18 +1,25 @@
 import streamlit as st
 import os
 import numpy as np
+from modules.database import supabase
 from modules.database import update_session_progress
 
-def show_intro(session_id):
+def show_intro():
     st.title("Experiment Description")
-    
-    # Load intro text from file
-    intro_file_path = os.path.join("assets", "text", "experiment_description.txt")
+
+    # Get the session State Scenario
+    scenario = st.session_state.get('scenario_id')
+    # Insert the scenario_id for the scenario "long" from the database
+    if scenario == '2e1e164a-699c-4c00-acff-61a98e23ddec' or 'b8426ff5-c6f2-4f25-a259-764e993ffa29':
+        intro_file_path = os.path.join("assets", "text", "100trial_experiment_description.txt")
+    else:
+        intro_file_path = os.path.join("assets", "text", "5trial_experiment_description.txt")
+
     with open(intro_file_path, "r", encoding="utf-8") as f:
         intro_text = f.read()
     st.write(intro_text)
-    
-    if st.button("Start Experiment"):
+
+    if st.button("Start Demo"):
         # Generate random demo data
         st.session_state.demo_data = {
             'ai_a': np.random.randint(0, 101),
@@ -20,20 +27,14 @@ def show_intro(session_id):
             'return_b': np.random.uniform(-0.1, 0.2),
         }
         st.session_state.demo_data['ai_b'] = 100 - st.session_state.demo_data['ai_a']
-        
-        # Update session state
-        st.session_state.update({
-            'page': 'demo',
-            'trial_step': 1,
-            'trial': 1
-        })
-        
-        # Update database
-        update_session_progress(
-            session_id,
-            page='demo',
-            trial=st.session_state.trial,
-            trial_step=st.session_state.trial_step
-        )
-        
-        st.rerun()
+        st.session_state.page = 'demo'
+        st.session_state.trial_step = 1
+
+        # Immediately update the DB to reflect "demo"
+        supabase.table('sessions').update({
+            'current_page': 'demo',
+            'current_trial': st.session_state.trial,
+            'current_trial_step': st.session_state.trial_step
+        }).eq('session_id', st.query_params['session_id']).execute()
+
+        st.rerun()
